refactor : package 명 변경 step3 -> RacingCar
